+++ uninstall.sh
@@ -1,7 +1,9 @@
 #!/bin/sh -e
 
 # remove the hackerschool into the site packages
-rm /usr/lib/python2.7/site-packages/hackerschool
+if [ -f /usr/lib/python2.7/site-packages/hackerschool ]; then
+    rm /usr/lib/python2.7/site-packages/hackerschool
+fi
 
 # add the server to start at boot
 sed -i '/start_server.sh/d' /etc/rc.local
